// src/types/auth.type.ts
export interface LoginBody {
  email: string;
  password: string;
  deviceId: string;
  fcmToken: string;
  platform: "android" | "ios";
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  age: number;
  deviceId: string;
  fcmToken: string;
  platform: "android" | "ios";
}

export interface RefreshBody {
  refreshToken: string;
}

export interface TokenResponse {
  accessToken: string;
  refreshToken: string;
}

export interface MessageResponse {
  message: string;
}